✒️Feat: 질문upvote/downvote,  답변 upvote/downvote 기능 구현

+++ back/src/main/java/team001_be/stackoverflowCloneDemo/answer/dto/AnswerResponseDto.java
@@ -12,7 +12,7 @@ public class AnswerResponseDto {
     private Long questionId;
     private Long userId;
     private String context;
-    private int voteCount;
+    private Long voteCount;
     private boolean isAccepted;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
